changed size_t to int

+++ CameraStreamShared/incglobal.h
@@ -63,8 +63,8 @@ char getch(){
 typedef struct _framehdr {
 	int frameDataType;
 	int frameDataTypeSz;
-	size_t frameRows = 0;
-	size_t frameCols = 0;
+	int frameRows = 0;
+	int frameCols = 0;
 } FrameHeader, *pFrameHeader;
 
 #pragma once
